+++ src/client/render.tsx
@@ -2,13 +2,13 @@
 import { h, render } from 'preact';
 import { HttpRegion } from 'httpyac';
 import type { NotebookRendererApi } from 'vscode-notebook-renderer';
-import { TestResults, RFC7230Response, HttpBody } from './components';
+import { TestResults, RFC7230Response, HttpBody, HljsMetaData } from './components';
 
 interface IRenderInfo {
   container: HTMLElement;
   mimeType: string;
   data: HttpRegion;
-  metaData?: Record<string, string>,
+  metaData?: HljsMetaData,
   notebookApi: NotebookRendererApi<unknown>;
 }
 
@@ -20,16 +20,16 @@ export function renderCell({ container, mimeType, data, metaData }: IRenderInfo)
   } else if (data.response) {
     switch (mimeType) {
       case 'x-application/httpbook-rfc7230':
-        render(<RFC7230Response response={data.response} rawBody={metaData?.rawBody} requestVisible={true} bodyVisible={true} />, container);
+        render(<RFC7230Response response={data.response} metaData={metaData} requestVisible={true} bodyVisible={true} />, container);
         break;
       case 'x-application/httpbook-rfc7230-response':
-        render(<RFC7230Response response={data.response} rawBody={metaData?.rawBody} bodyVisible={true} />, container);
+        render(<RFC7230Response response={data.response} metaData={metaData} bodyVisible={true} />, container);
         break;
       case 'x-application/httpbook-rfc7230-header':
-        render(<RFC7230Response response={data.response} rawBody={metaData?.rawBody} requestVisible={true} />, container);
+        render(<RFC7230Response response={data.response} metaData={metaData} requestVisible={true} />, container);
         break;
       default:
-        render(<HttpBody body={data.response.body} rawBody={metaData?.rawBody} mimeType={data.response.contentType?.mimeType} />, container);
+        render(<HttpBody response={data.response} metaData={metaData} />, container);
         break;
     }